jobs-for-reprocessing-tables/s3-to-postgres/requirements.txt
boto3==1.19.9
botocore==1.22.9
certifi==2021.10.8
charset-normalizer==2.0.7
docopt==0.6.2
idna==3.3
Jinja2==3.0.2
jinjasql==0.1.8
jmespath==0.10.0
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
smart-open==5.2.1
urllib3==1.26.7
yarg==0.1.9
